fix: resolve remaining React lint errors
