+++ packages/service/client/src/pages/Runner.tsx
@@ -18,6 +18,12 @@ import { useTheme } from '@/lib/theme';
 
 const REFRESH_INTERVAL = 10_000;
 
+/** Compute key age in days from a creation timestamp. */
+function computeKeyAge(keyCreatedAt: string | null | undefined): string | null {
+  if (!keyCreatedAt) return null;
+  return `${Math.floor((Date.now() - new Date(keyCreatedAt).getTime()) / 86_400_000)}d`;
+}
+
 export function Runner() {
   const [jobs, setJobs] = useState<RunnerJob[]>([]);
   const [stats, setStats] = useState<RunnerStats | null>(null);
@@ -29,10 +35,11 @@ export function Runner() {
   const [theme, toggleTheme] = useTheme();
   const { isInsider, searchEnabled, keyCreatedAt, rotateKey } = useAuth();
 
-  // Compute key age string
-  const keyAge = keyCreatedAt
-    ? `${Math.floor((Date.now() - new Date(keyCreatedAt).getTime()) / 86_400_000)}d`
-    : null;
+  // Compute key age string (in effect to avoid impure Date.now during render)
+  const [keyAge, setKeyAge] = useState<string | null>(null);
+  useEffect(() => {
+    queueMicrotask(() => { setKeyAge(computeKeyAge(keyCreatedAt)); });
+  }, [keyCreatedAt]);
 
   const fetchData = useCallback(async () => {
     try {
@@ -51,7 +58,7 @@ export function Runner() {
   }, []);
 
   useEffect(() => {
-    void fetchData();
+    queueMicrotask(() => { void fetchData(); });
   }, [fetchData]);
 
   useEffect(() => {
